refactor[Provider, Role]: move to Enum folder
