+++ src/main/java/com/example/demo/oauth/UserProfile.java
@@ -1,8 +1,8 @@
 package com.example.demo.oauth;
 
-import com.example.demo.user.domain.entity.Provider;
-import com.example.demo.user.domain.entity.Role;
 import com.example.demo.user.domain.entity.Users;
+import com.example.demo.user.domain.enums.Provider;
+import com.example.demo.user.domain.enums.Role;
 import lombok.Builder;
 import lombok.Getter;
