+++ src/main/java/com/example/demo/user/domain/enums/Provider.java
@@ -1,6 +1,7 @@
-package com.example.demo.user.domain.entity;
+package com.example.demo.user.domain.enums;
 
 public enum Provider {
     GOOGLE,KAKAO,NAVER
 
 }
+
